git-svn-id: https://repo.it.itba.edu.ar/svn/sia-2012-4@35 39a6b423-ae9d-42ef-8f9a-67f0a12ae516

+++ TP01/src/com/g7/rules/DropThree.java
@@ -24,7 +24,7 @@ public class DropThree implements GPSRule{
 	}
 
 	@Override
-	public Integer getCost() {
+	public float getCost() {
 		return 1;
 	}
