+++ TP01/src/com/g7/rules/SevenToThree.java
@@ -26,7 +26,7 @@ public class SevenToThree implements GPSRule{
 	}
 
 	@Override
-	public Integer getCost() {
+	public float getCost() {
 		return 1;
 	}
