+++ TP01/src/com/g7/rules/TenToSeven.java
@@ -26,7 +26,7 @@ public class TenToSeven implements GPSRule {
 	}
 
 	@Override
-	public Integer getCost() {
+	public float getCost() {
 		return 1;
 	}
