+++ TP01/src/gps/api/GPSProblem.java
@@ -32,6 +32,6 @@ public interface GPSProblem {
 	 * @param state The state where the Heuristic should be computed.
 	 * @return The value of the Heuristic.
 	 */
-	Integer getHValue(GPSState state);
+	float getHValue(GPSState state);
 
 }
